+++ auth.js
@@ -2,16 +2,18 @@ const jwt = require("jsonwebtoken");
 var secret = "This is the secret for signing tokens";
 
 module.exports = function (req, res, next) {
-  token = req.cookies.token;
-  if (!token) {
-    return res.status(400).json({ message: "not authorized" });
-  }
+  var authheader = req.headers.authorization;
 
-  try {
-    const decoded = jwt.verify(token, secret);
-    req.user = decoded;
-    next();
-  } catch (ex) {
-    res.status(400).send("Invalid JWT.");
+  if (!authheader) {
+    res.json({ message: "This is not an auth route" });
+  } else {
+    const token = authheader.split(" ")[1];
+
+    try {
+      jwt.verify(token, secret);
+      next();
+    } catch (err) {
+      res.json({ message: "token doesn't match" });
+    }
   }
 };
